Create Daily Disk Space Report
This batch script copied and pasted a copy of our spreadsheet template into a destination directory.  The script then renamed the file to match the other reports, and added the current date to the filename. 

Finally, the script opens the file so that I can input the data that was produced from an SQL report that queries our server's drives for disk space under a certain threshold.

These reports were created daily and I wrote this script to cut out some technical rigor involved in the process.